Took away gun drawing, looks like shit.

+++ src/Zapper.hpp
@@ -6,7 +6,7 @@
 class Zapper {
 protected:
 	GLfloat location[4][2];	// Original location of the gun
-	GLfloat rotation; // Current rotation to point at mouse
+	GLint rotation; // Current rotation to point at mouse, 0 = left, 1 = center, 2 = right
 public:
 	Zapper(); // Zapper constructor
 	void draw(); // Drawing the Zapper
